fix: improve code readability and add debug options
- Refactored `GetBuildDepthOnce` to use `static_cast<int>`
- Simplified `std::cout` statements in `validation.hpp`
- Corrected `assert` condition in `splitter.h`
- Added `set -o xtrace` to `checkCorrect.sh` for debugging
- Updated `test_framework.h` to use semicolons in `PrintTreeParam`

+++ include/pstp/base_tree.h
@@ -348,7 +348,9 @@ class BaseTree {
 
   Box GetRootBox() { return this->tree_box_; }
 
-  consteval static auto GetBuildDepthOnce() { return kBuildDepthOnce; }
+  consteval static auto GetBuildDepthOnce() {
+    return static_cast<int>(kBuildDepthOnce);
+  }
 
  protected:
   Node* root_ = nullptr;
